Issue #1, Issue#2 and Issue#3
fixed #1 : xyzshiftplus and xyzshiftminus are no longer generated for sorts not
containing variable bindings.

fixed #2 : type signatures are now also generated, both for the Haskell and OCaml output.
The Type signatures are internally represented in a more global format so that converting
that format to either Haskell or OCaml was easier.

fixed #3 : Documentation is also generated along with the functions.