Bug fixed, updated generated code from specifications
There was a bug in the typesignature generation when multiple namespaces where involved.
It should now correctly generate the typesignatures.

The generated code from the specifications were updated in the corresponding calculus folder.
In the case of the ExEff calculus, the ExEffImpl code also needed some small changes because
extra code was added to the previous generated ExEff.hs code. That extra code does not need to
be added anymore so that newly generated code can easily replace the old one now.